Move date validator to format.swift
